Moved various quantity tests into test_quantities
